Revert "fix(migrations): ship resource links under 1.9.6, not 2.0.0"
CE main deliberately renamed the next release from 1.9.6 to 2.0.0
(923e413358, 2026-07-02), so 1.9.6 will never ship. The feature stays
folded in V25 keyed to 2.0.0 with the V27 filters gated at 2.0.0.
