+++ src/Appwrite/Migration/Migration.php
@@ -98,7 +98,6 @@ abstract class Migration
         '1.9.3' => 'V24',
         '1.9.4' => 'V24',
         '1.9.5' => 'V24',
-        '1.9.6' => 'V26',
         '2.0.0' => 'V25',
     ];
